feat: CategoryHeader 컴포넌트 구현
Resolves: #29
